dc-back-end/dc-base-service/src/main/java/tech/kjworks/dcbaseservice/util/exception/Error.java
package tech.kjworks.dcbaseservice.util.exception;

public class Error {
    private String message;
    private Integer status;
    private Long timestamp;

    public Error() {}

    public Error(String message,
        Integer status,
        Long timestamp) {
        this.message = message;
        this.status = status;
        this.timestamp = timestamp;
    }

    public String getMessage() {
        return message;
    }

    public void setMessage(String message) {
        this.message = message;
    }

    public Integer getStatus() {
        return status;
    }

    public void setStatus(Integer status) {
        this.status = status;
    }

    public Long getTimestamp() {
        return timestamp;
    }

    public void setTimestamp(Long timestamp) {
        this.timestamp = timestamp;
    }
}
